requirements.txt
faster-whisper
pyautogui
pytesseract
opencv-python
sounddevice
numpy
Pillow
